Feat: add service unit tests and correct class dictionary

+++ backend/app/schemas/store_schema.py
@@ -2,4 +2,4 @@ from pydantic import BaseModel
 
 class Pair(BaseModel):
     key: str
-    value: int
+    value: float
